test/chaos-pipeline/apps/consecutive_create_and_update_operations/requirements.txt
weaviate-client==3.3.2
